backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy>=2.0
psycopg[binary]>=3.2
alembic>=1.13
pydantic>=2.7
pydantic-settings>=2.3
python-jose[cryptography]>=3.3
pwdlib[argon2]>=0.2
python-multipart>=0.0.9
email-validator>=2.1
pytest>=8.0
httpx>=0.27
pytest-cov>=5.0
